Moving get next direction methods to the piece class
