+++ edu/colostate/cs/cs414/StringCheese/src/ChessBoard.java
@@ -38,6 +38,7 @@ public class ChessBoard {
 
     }
 
+
     public ChessPiece getPiece(String position) throws IllegalPositionException {
         // This method returns the chess piece at a given position. The position is represented as a two-character
         // string (e.g., e8) as described above. The first letter is in lowercase (a..h) and the second letter is a
@@ -45,6 +46,10 @@ public class ChessBoard {
         // position outside the board, the exception is thrown.
         checkValidPosition(position);
         return board[getRow(position)][getCol(position)];
+    }
+
+    public ChessPiece getPiece(int row, int col){
+        return board[row][col];
     }
 
     public HashSet<String> selectPiece(String position){
@@ -55,7 +60,7 @@ public class ChessBoard {
             if(piece == null){
                 return moves;
             }
-            moves = piece.getValidMoves();
+            moves = piece.legalMoves();
             return moves;
         } catch (IllegalPositionException e) {
             return moves;
@@ -86,11 +91,11 @@ public class ChessBoard {
         }
     }
  
-    private int getRow(String position) {
+    public int getRow(String position) {
         return position.charAt(0) - 'a';
     }
 
-    private int getCol(String position){
+    public int getCol(String position){
         return Character.getNumericValue(position.charAt(1)) - 1;
     }
 
@@ -120,231 +125,7 @@ public class ChessBoard {
             throw new IllegalMoveException("Cannot move to position " + toPosition);
         }
     }
-	
-	 public HashSet<String> getNextStraightClockwise(String position) throws IllegalPositionException {
-        int index;
-        String newPos;
-        if(innerRing.contains(position)){
-            index = innerRing.indexOf(position);
-            //circular array, so next of .size() is index 0
-            if(index+1 > innerRing.size()){
-                newPos = innerRing.get(0);
-            }else {
-                newPos = innerRing.get(index + 1);
-            }
-        }else if(outerRing.contains(position)){
-            index = outerRing.indexOf(position);
-            if(index+1 > outerRing.size()){
-                newPos = outerRing.get(0);
-            }else {
-                newPos = outerRing.get(index + 1);
-            }
-        }else {
-            throw new IllegalPositionException("Position given not valid");
-        }
-        int row = getRow(newPos);
-        int col = getCol(newPos);
-        int oldRow = getRow(position);
-        int oldCol = getCol(position);
-        HashSet<String> moves = new HashSet<>();
-        if(board[row][col] != null && board[row][col].color == board[oldRow][oldCol].color){
-            return moves;
-        }else{
-            moves.add(newPos);
-            return moves;
-        }
-    }
 
-    public HashSet<String> getBackwardStraight(String position) throws IllegalPositionException {
-        int index;
-        String newPos;
-        if(innerRing.contains(position)){
-            index = innerRing.indexOf(position);
-            //circular array, so prev of index(0) is .size()-1
-            if(index-1 < 0){
-                newPos = innerRing.get(innerRing.size()-1);
-            }else {
-                newPos = innerRing.get(index - 1);
-            }
-        }else if(outerRing.contains(position)){
-            index = outerRing.indexOf(position);
-            if(index-1 < 0){
-                newPos = outerRing.get(innerRing.size()-1);
-            }else {
-                newPos = outerRing.get(index - 1);
-            }
-        }else {
-            throw new IllegalPositionException("Position given not valid");
-        }
-        int row = getRow(newPos);
-        int col = getCol(newPos);
-        int oldRow = getRow(position);
-        int oldCol = getCol(position);
-        HashSet<String> moves = new HashSet<>();
-        if(board[row][col] != null && board[row][col].color == board[oldRow][oldCol].color){
-            return new HashSet<String>();
-        }else{
-            return new HashSet<String>(Arrays.asList(newPos));
-        }
-    }
-
-
-
-    /*
-         HashSet<String> move = new HashSet<>();
-        int row = getRow(position);
-        int col = getCol(position);
-
-        if (row - 1 >= 0 && col - 1 >= 0) {
-            move.add(board[row - 1][col - 1].getPosition());
-        }
-        if (row + 1 <= board.length - 1 && col - 1 >= 0) {
-            move.add(board[row + 1][col - 1].getPosition());
-        }
-        if (row + 1 <= board.length - 1 && col + 1 <= board.length - 1) {
-            move.add(board[row + 1][col + 1].getPosition());
-        }
-        if (row - 1 >= 0 && col + 1 <= board.length - 1) {
-            move.add(board[row - 1][col + 1].getPosition());
-        }
-        return move;
-
-     */
-
-    public HashSet<String> getNextDiagonalClockwise(String position) {
-        HashSet<String> legalMoves = new HashSet<>();
-        char letter = position.charAt(0);
-        int row = getRow(position);
-        //if ( row <=2 and col > b ) move left i.e. decrease col by 1 and row can be 1 or 2. with exception if c2 then also b3
-        if(row <=1 && letter > 'b'){
-            letter -= 1;
-            if(row == 1) legalMoves.add(Character.toString(letter).concat(Integer.toString(0)));
-            else legalMoves.add(Character.toString(letter).concat(Integer.toString(1)));
-        }
-        //if ( col <=b and row < 5 ) move up i.e. increase row by 1 and col can be a or b. with exception if b5 then also c6
-        if(letter <= 'b' && row < 5){
-            //don't need?
-                if(letter == 'a') legalMoves.add(Character.toString('b').concat(Integer.toString(row+1)));
-                else legalMoves.add(Character.toString('a').concat(Integer.toString(row+1)));
-        }
-        //if ( row >=5 and col < f ) move right i.e. increase col by 1 and row can be 6 or 7. with exception if e6 then also f5
-        if(row >=5 && letter < 'f'){
-            letter += 1;
-            if(row == 5) legalMoves.add(Character.toString(letter).concat(Integer.toString(6)));
-            else legalMoves.add(Character.toString(letter).concat(Integer.toString(5)));
-        }
-        //if ( col >=f and row > 1 ) move down i.e. decrease row by 1 and col can be f or g. with exception f3 can move e2
-        if(row >1 && letter >= 'f'){
-            if(letter == 'f')  legalMoves.add(Character.toString('g').concat(Integer.toString(row-1)));
-            else legalMoves.add(Character.toString('f').concat(Integer.toString(row-1)));
-        }
-
-        //special cases with 3 possible moves
-        if(position.equals("c2")){
-            legalMoves.add("b3");
-        }
-        if(position.equals("b5")){
-            legalMoves.add("c6");
-        }
-        if(position.equals("e6")){
-            legalMoves.add("f5");
-        }
-        if(position.equals("f3")){
-            legalMoves.add("e2");
-        }
-        return legalMoves;
-    }
-
-    public HashSet<String> getBackwardDiagonals(String position) {
-        HashSet<String> legalMoves = new HashSet<>();
-        char letter = position.charAt(0);
-        int row = getRow(position);
-        //if ( row <=2 and col > b ) move left i.e. decrease col by 1 and row can be 1 or 2. with exception if c2 then also b3
-        if(row <=1 && letter < 'f'){
-            letter += 1;
-            if(row == 1) legalMoves.add(Character.toString(letter).concat(Integer.toString(0)));
-            else legalMoves.add(Character.toString(letter).concat(Integer.toString(1)));
-        }
-        //if ( col <=b and row < 5 ) move up i.e. increase row by 1 and col can be a or b. with exception if b5 then also c6
-        if(letter <= 'b' && row > 2 ){
-            if(letter == 'a') legalMoves.add(Character.toString('b').concat(Integer.toString(row-1)));
-            else legalMoves.add(Character.toString('a').concat(Integer.toString(row-1)));
-        }
-        //if ( row >=5 and col < f ) move right i.e. increase col by 1 and row can be 6 or 7. with exception if e6 then also f5
-        if(row >=5 && letter > 'b'){
-            letter -= 1;
-            if(row == 5) legalMoves.add(Character.toString(letter).concat(Integer.toString(6)));
-            else legalMoves.add(Character.toString(letter).concat(Integer.toString(5)));
-        }
-        //if ( col >=f and row > 1 ) move down i.e. decrease row by 1 and col can be f or g. with exception f3 can move e2
-        if(row <= 4 && letter >= 'f'){
-            if(letter == 'f')  legalMoves.add(Character.toString('g').concat(Integer.toString(row+1)));
-            else legalMoves.add(Character.toString('f').concat(Integer.toString(row+1)));
-        }
-
-        //special cases with 3 possible moves
-        if(position.equals("c2")){
-            legalMoves.add("b3");
-        }
-        if(position.equals("b5")){
-            legalMoves.add("c6");
-        }
-        if(position.equals("e6")){
-            legalMoves.add("f5");
-        }
-        if(position.equals("f3")){
-            legalMoves.add("e2");
-        }
-        return legalMoves;
-
-        //return new HashSet<String>();
-    }
-
-    //returns adjacent tiles on opposite ring
-    public HashSet<String> getSideways(String position) {
-        /*
-        //if on innerRing corners add straight on outer ring
-        {
-            if (position.equals("b2") && (board[getRow("a2")][getCol("a2")] != null && board[getRow("a2")][getCol("a2")].color != board[oldRow][oldCol].color)) {
-                moves.add("a2");
-            }
-            if (position.equals("b6") && (board[getRow("b7")][getCol("b7")] != null && board[getRow("b7")][getCol("b7")].color != board[oldRow][oldCol].color)) {
-                moves.add("b7");
-            }
-            if (position.equals("f6") && (board[getRow("g6")][getCol("g6")] != null && board[getRow("g6")][getCol("g6")].color != board[oldRow][oldCol].color)) {
-                moves.add("g6");
-            }
-            if (position.equals("f2") && (board[getRow("f1")][getCol("f1")] != null && board[getRow("f1")][getCol("f1")].color != board[oldRow][oldCol].color)) {
-                moves.add("f1");
-            }
-        }
-
-
-        //add straight on opposite ring
-        if(position.equals("a2") && (board[getRow("b2")][getCol("b2")] != null && board[getRow("b2")][getCol("b2")].color != board[oldRow][oldCol].color)) {
-            moves.add("b2");
-        }
-        if(position.equals("b7") && (board[getRow("b6")][getCol("b6")] != null && board[getRow("b6")][getCol("b6")].color != board[oldRow][oldCol].color)) {
-            moves.add("b6");
-        }
-        if(position.equals("g6") && (board[getRow("f6")][getCol("f6")] != null && board[getRow("f6")][getCol("f6")].color != board[oldRow][oldCol].color)) {
-            moves.add("f6");
-        }
-        if(position.equals("f1") && (board[getRow("f2")][getCol("f2")] != null && board[getRow("f2")][getCol("f2")].color != board[oldRow][oldCol].color)) {
-            moves.add("f2");
-        }
-        */
-
-
-
-        return new HashSet<String>();
-    }
-
-
-
-	
-	
-	
 
     public String toString() {
         String s = "";
